+++ src/extension/content.ts
@@ -2,9 +2,13 @@ import { log } from "./log";
 
 import { fetchStream, streamBody } from "./fetchStream";
 
-import { MR_WIDGET_SECTION,GITLAB_CLI_DESC,GITLAB_CLI_BUTTON } from './constants/domClasses';
+import {
+  MR_WIDGET_SECTION,
+  GITLAB_CLI_DESC,
+  GITLAB_CLI_BUTTON,
+} from "./constants/domClasses";
 
-import {getMergeRequestInfo} from "./api";
+import { getMergeRequestInfo } from "./api";
 
 function renderButton() {
   const button = document.createElement("button");
@@ -37,6 +41,8 @@ function renderMergeButton(sourceBranch, targetBranch) {
         }&source=${encodeURIComponent(
           sourceBranch!
         )}&target=${encodeURIComponent(targetBranch!)}`,
+        "GET",
+        null,
         (chunkString) => {
           button.textContent = "Merging";
           setContentInDesc(chunkString);
@@ -54,7 +60,6 @@ function renderMergeButton(sourceBranch, targetBranch) {
   buttonGroup.appendChild(button);
 }
 
-
 function renderDescription() {
   const descriptionAreaEl = document.createElement("p");
   descriptionAreaEl.id = GITLAB_CLI_DESC;
@@ -71,13 +76,13 @@ function renderDescription() {
 
 function render() {
   const rootDiv = document.createElement("div");
-  rootDiv.style.display='flex';
-  rootDiv.style.flexDirection='column';
-  rootDiv.style.marginLeft='auto';
-  rootDiv.style.marginRight='auto';
-  rootDiv.style.marginTop='16px';
+  rootDiv.style.display = "flex";
+  rootDiv.style.flexDirection = "column";
+  rootDiv.style.marginLeft = "auto";
+  rootDiv.style.marginRight = "auto";
+  rootDiv.style.marginTop = "16px";
   rootDiv.classList.add("mr-widget-heading", "append-bottom-default");
-  rootDiv.style.border='none';
+  rootDiv.style.border = "none";
 
   const containerDiv = document.createElement("div");
   containerDiv.classList.add("mr-widget-content");
@@ -92,9 +97,8 @@ function render() {
   return rootDiv;
 }
 
-
 function getButtons() {
-  return document.querySelectorAll("."+GITLAB_CLI_BUTTON);
+  return document.querySelectorAll("." + GITLAB_CLI_BUTTON);
 }
 
 function disableButtons() {
@@ -112,95 +116,111 @@ function wait(millis) {
   return new Promise((res) => setTimeout(res, millis));
 }
 
-async function initialise(repoURLName,mergeRequestID,sourceBranch,targetBranch,isRebaseInProgress) {
+async function initialise(
+  repoURLName,
+  mergeRequestID,
+  sourceBranch,
+  targetBranch,
+  isRebaseInProgress
+) {
   const referenceEl = document.querySelector(MR_WIDGET_SECTION);
   const el = render();
   referenceEl.classList.add("mr-widget-workflow");
   referenceEl.prepend(el);
 
-  renderMergeButton(sourceBranch,targetBranch);
-  const mergeButton = document.getElementById('gitlab-cli-merge');
+  renderMergeButton(sourceBranch, targetBranch);
+  const mergeButton = document.getElementById("gitlab-cli-merge");
   disableButtons();
   mergeButton.classList.remove(GITLAB_CLI_BUTTON);
-  try{
-    await fetch(`http://localhost:4000/handshake?location=${window.location}`)
-    .then((r) => {
-    if (r.status === 200) {
-      const mergeButton = document.getElementById('gitlab-cli-merge');
-      mergeButton.classList.add(GITLAB_CLI_BUTTON);
-      enableButtons();
-      return true;
-    }
-    if (r.status === 500) {
-      return false;
-    }
-    if (r.status === 512) {
-      const descEl = document.getElementById(GITLAB_CLI_DESC);
-      setContentInDesc("CLI busy");
-      streamBody(r.body, (chunkString) => {
-        descEl.textContent = chunkString;
-      }).then(() => {
-        clearContentInDesc();
-        const mergeButton = document.getElementById('gitlab-cli-merge');
+  try {
+    await fetch(
+      `http://localhost:4000/handshake?location=${window.location}`
+    ).then((r) => {
+      if (r.status === 200) {
+        const mergeButton = document.getElementById("gitlab-cli-merge");
         mergeButton.classList.add(GITLAB_CLI_BUTTON);
         enableButtons();
-      });
+        return true;
+      }
+      if (r.status === 500) {
+        return false;
+      }
+      if (r.status === 512) {
+        const descEl = document.getElementById(GITLAB_CLI_DESC);
+        setContentInDesc("CLI busy");
+        streamBody(r.body, (chunkString) => {
+          descEl.textContent = chunkString;
+        }).then(() => {
+          clearContentInDesc();
+          const mergeButton = document.getElementById("gitlab-cli-merge");
+          mergeButton.classList.add(GITLAB_CLI_BUTTON);
+          enableButtons();
+        });
+        return false;
+      }
       return false;
-    }
-    return false;
     });
-  }catch{
-    console.log('Server not initialised!');
+  } catch {
+    console.log("Server not initialised!");
     setContentInDesc(`Server not initialised`);
   }
 }
 
-function getProjectInfo(pathName){
+function getProjectInfo(pathName) {
   let pathArray = pathName.split("/");
-  const midIndex = pathArray.findIndex((element)=>{
-    return element == 'merge_requests';
-  })
-  const repoURLIndex = pathArray.slice(1,midIndex-1);
-  const repoURLName = repoURLIndex.join('/');
-  return { mergeRequestID: pathArray.at(midIndex+1), repoURLName};
+  const midIndex = pathArray.findIndex((element) => {
+    return element == "merge_requests";
+  });
+  const repoURLIndex = pathArray.slice(1, midIndex - 1);
+  const repoURLName = repoURLIndex.join("/");
+  return { mergeRequestID: pathArray.at(midIndex + 1), repoURLName };
 }
-async function renderWidget(projectInfo){
+async function renderWidget(projectInfo) {
   let retryCounter = 1;
-  while(retryCounter<=2){
-    try{
-      let res = await getMergeRequestInfo(projectInfo.repoURLName,projectInfo.mergeRequestID);
-      if(!res.isMerged){
-        initialise(projectInfo.repoURLName,projectInfo.mergeRequestID,res.sourceBranch,res.targetBranch,res.isRebaseInProgress);
+  while (retryCounter <= 2) {
+    try {
+      let res = await getMergeRequestInfo(
+        projectInfo.repoURLName,
+        projectInfo.mergeRequestID
+      );
+      if (!res.isMerged) {
+        initialise(
+          projectInfo.repoURLName,
+          projectInfo.mergeRequestID,
+          res.sourceBranch,
+          res.targetBranch,
+          res.isRebaseInProgress
+        );
         return;
       }
-    }catch(e){
-        console.log(e);
+    } catch (e) {
+      console.log(e);
     }
     await wait(2000);
-    retryCounter+=1;
+    retryCounter += 1;
   }
 }
 const main = () => {
   log("init");
   const pathName = window.location.pathname;
   const projectInfo = getProjectInfo(pathName);
-  const sectionContainer = document.querySelector('.mr-section-container');
-  if(sectionContainer!=null){
+  const sectionContainer = document.querySelector(".mr-section-container");
+  if (sectionContainer != null) {
     renderWidget(projectInfo);
     return;
   }
-  const targetNode = document.querySelector('.issuable-discussion');
+  const targetNode = document.querySelector(".issuable-discussion");
   const config = { childList: true, subtree: true };
 
-  const callback = function(mutationList, observer) {
-  for(const mutation of mutationList) {
-      if(mutation.target.classList.contains('mr-section-container')){
-        log('Widget section loaded');
+  const callback = function (mutationList, observer) {
+    for (const mutation of mutationList) {
+      if (mutation.target.classList.contains("mr-section-container")) {
+        log("Widget section loaded");
         observer.disconnect();
         renderWidget(projectInfo);
         break;
       }
-  }
+    }
   };
 
   const observer = new MutationObserver(callback);
