+++ src/extension/fetchStream.ts
@@ -19,7 +19,7 @@ function streamBody(body, onChunkReceive) {
             // Enqueue the next data chunk into our target stream
             controller.enqueue(value);
             const chunkString = decoder.decode(value, { stream: true });
-            if (chunkString.toLowerCase().startsWith('error')) {
+            if (chunkString.toLowerCase().startsWith("error")) {
               onChunkReceive(chunkString);
               throw Error(chunkString);
             }
@@ -33,11 +33,26 @@ function streamBody(body, onChunkReceive) {
       });
     })
     .then((rs) => new Response(rs))
-    .then((response) => response.text())
+    .then((response) => response.text());
 }
 
-function fetchStream(url, onChunkReceive) {
-  return fetch(url)
+function fetchBuilder(url, method, payload) {
+  if (method === "GET") {
+    return fetch(url);
+  } else if (method === "POST") {
+    return fetch(url, {
+      method: "POST",
+      mode: "cors",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(payload),
+    });
+  }
+}
+
+function fetchStream(url, method, payload, onChunkReceive) {
+  return fetchBuilder(url, method, payload)
     .then((r) => {
       if (r.status >= 400) {
         return r.text().then((text) => {
@@ -45,7 +60,8 @@ function fetchStream(url, onChunkReceive) {
         });
       }
       return r.body;
-    }).catch(e=>{
+    })
+    .catch((e) => {
       onChunkReceive(e);
       throw e;
     })
