+++ src/extension/public/cherrypick.html
@@ -8,7 +8,7 @@
       crossorigin="anonymous"
     />
     <link rel="stylesheet" href="style.css" />
-    <script src="script.js" defer></script>
+    <script src="assets/script.js" defer></script>
   </head>
   <body>
     <div class="app-header">
