+++ src/extension/public/index.html
@@ -23,7 +23,7 @@
         margin:10px;
       }
     </style>
-    <script src="popup.js" defer></script>
+    <script src="assets/popup.js" defer></script>
   </head>
   <body>
     <div class="app-header">
