+++ src/extension/utils.ts
@@ -1,0 +1,6 @@
+async function getCurrentTab() {
+    const queryOptions = { active: true, lastFocusedWindow: true };
+    const [tab] = await chrome.tabs.query(queryOptions);
+    return tab;
+}
+export {getCurrentTab};
